requirements.txt
streamlit>=1.20
pycryptodome>=3.17